Add NodeApplicationLifecycleController and use that instead of NodeMainPresenter for app services and tor lifecycle.
NodeMainPresenter lifecycle is tied to activity but we want our core services and tor to be bound to the application/process lifecycle.
We want to hava guarantee that those services are initialized only once.

As Android does not provide an event when the app gets killed we cannot do a graceful shutdown. Follow up commits might address that issue, but we need to see in tests if that actually cause a problem. The risks on the core services is limited and for tor it also seems that hard shutdown of tor does not cause problems.
The best solution might be a foreground service but the added complexity might not justify the uncertain benefits. And it also does not provide a guarantee to enable graceful shutdown handling in all cases.
